pyproject.toml
[tool.poetry]
name = "privacychatslm"
version = "0.1.0"
description = ""
authors = ["DiegoSamoo <diego.sanchez.moreno@alumnos.upm.es>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
configparser = "^7.1.0"
langgraph = "^0.2.53"
langgraph-sdk = "^0.1.40"
langgraph-checkpoint-sqlite = "^2.0.1"
langsmith = "^0.1.147"
langchain-community = "^0.3.8"
langchain-core = "^0.3.21"
langchain-openai = "^0.2.10"
notebook = "^7.2.2"
tavily-python = "^0.5.0"
wikipedia = "^1.4.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
isort = "^5.13.2"
langchain-ollama = "^0.2.0"
pillow = "^11.0.0"
presidio-analyzer = {extras = ["transformers"], version = "^2.2.357"}
presidio-anonymizer = "^2.2.357"
huggingface-hub = "^0.28.1"
transformers = "^4.48.2"
torch = "^2.6.0"
numpy = "1.26.4"
thinc = "8.2.2"
spacy = "3.7.4"
ollama = "^0.4.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
